backend/requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.1.8
coverage==7.10.7
cryptography==46.0.7
Deprecated==1.3.1
exceptiongroup==1.3.1
fastapi==0.115.6
google-auth==2.49.2
google-genai==1.5.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.13
iniconfig==2.1.0
limits==4.2
packaging==24.2
pluggy==1.6.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.13.3
pydantic-settings==2.7.1
pydantic_core==2.46.3
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
python-dotenv==1.0.1
PyYAML==6.0.3
requests==2.32.5
slowapi==0.1.9
starlette==0.41.3
tomli==2.4.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.34.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==14.2
wrapt==2.1.2
